==> Scripts/1_data_extraction.ipynb <==
import ijson
import pandas as pd

with open('all_data/swiggy.json', 'r') as f:
    # Use 'Bangalore.restaurants' if the JSON top-level keys are city names.
    restaurants_iterator = ijson.kvitems(f, 'Bangalore')
    bangalore_data = list(restaurants_iterator)

# Initialize an empty list to hold restaurant data
restaurants_list = []

# Iterate over the data to extract restaurant details
for area, area_data in bangalore_data:
    for rest_id, rest_data in area_data['restaurants'].items():
        # Flatten the menu into a string representation
        menu_items = []
        for category, items in rest_data.get('menu', {}).items():
            for item_name, item_details in items.items():
                menu_items.append(f"{item_name} ({item_details['veg_or_non_veg']}): ₹{item_details['price']}")
        menu_str = "; ".join(menu_items)
        
        # Append the restaurant data to the list
        restaurants_list.append({
            'Area': area,
            'Restaurant ID': rest_id,
            'Name': rest_data.get('name', ''),
            'Rating': rest_data.get('rating', ''),
            'Rating Count': rest_data.get('rating_count', ''),
            'Cost for Two': rest_data.get('cost', ''),
            'Address': rest_data.get('address', ''),
            'Cuisine': rest_data.get('cuisine', ''),
            'License No': rest_data.get('lic_no', ''),
            'Menu': menu_str,
            'Link': rest_data.get('link', '')
        })

# Create a DataFrame
df = pd.DataFrame(restaurants_list)

# Display the DataFrame
print(df)

df.to_csv('all_data/swiggy_bangalore.csv', index=False)

df_indian_restaurants = pd.read_csv('all_data/indian_restaurants.csv')
df_bangalore_restaurants = df_indian_restaurants[df_indian_restaurants['City'] == 'Bangalore']
df_bangalore_restaurants.to_csv('all_data/indian_restaurants_bangalore.csv', index=False)
