==> Scripts/2_data_processing.ipynb <==
import pandas as pd
# pd.set_option('display.max_colwidth', None)
# pd.reset_option('display.max_colwidth')

# LOADING DATA

swiggy_df = pd.read_csv('all_data/swiggy_bangalore.csv')
zomato_df = pd.read_csv('all_data/zomato_bangalore.csv')
indian_restaurants_df = pd.read_csv('all_data/indian_restaurants_bangalore.csv')

zomato_df.head()

zomato_columns_to_keep = ['name', 'address', 'location', 'cuisines', 'rate','votes', 'approx_cost(for two people)']
zomato_df = zomato_df[zomato_columns_to_keep]
zomato_df.head()

nan_counts = zomato_df.isna().sum()
print(nan_counts)

zomato_df.describe(include='all')

swiggy_df.head()

swiggy_columns_to_keep = ['Name', 'Address', 'Area', 'Cuisine', 'Rating', 'Rating Count', 'Cost for Two']
swiggy_df = swiggy_df[swiggy_columns_to_keep]
swiggy_df.head()

swiggy_df['Rating Count'] = swiggy_df['Rating Count'].str[:-9]
swiggy_df['Rating Count'] = swiggy_df['Rating Count'].str.replace('K', '000')
swiggy_df['Rating Count'] = swiggy_df['Rating Count'].str.replace('Too Fe', '10')
swiggy_df['Rating Count'].drop_duplicates()

indian_restaurants_df.head()

indian_restaurants_columns_to_keep = ['Name', 'Location', 'Locality', 'Cuisine','Rating', 'Votes', 'Cost']
indian_restaurants_df = indian_restaurants_df[indian_restaurants_columns_to_keep]
indian_restaurants_df.head()

indian_restaurants_df['Locality'].drop_duplicates().head(50)

print(f"Swiggy:{swiggy_df.columns}")
print(f"Zomato:{zomato_df.columns}")
print(f"Indian Restaurants:{indian_restaurants_df.columns}")

# Standardize column names across all dataframes to a common format
swiggy_df.columns = ['name', 'address', 'location', 'cuisines', 'rating', 'rating_count', 'cost_for_two']
zomato_df.columns = ['name', 'address', 'location', 'cuisines', 'rating', 'rating_count', 'cost_for_two']
indian_restaurants_df.columns = ['name', 'address', 'location', 'cuisines', 'rating', 'rating_count', 'cost_for_two']

# Display the updated column names
print(f"Swiggy: {swiggy_df.columns}")
print(f"Zomato: {zomato_df.columns}")
print(f"Indian Restaurants: {indian_restaurants_df.columns}")

zomato_df['rating'] = zomato_df['rating'].str.replace('/5', '')
zomato_df

print(f"Swiggy:{swiggy_df.shape}")
print(f"Zomato:{zomato_df.shape}")
print(f"Indian Restaurants:{indian_restaurants_df.shape}")

# count of nulls in each dataframe each column

print(f"Swiggy:\n{swiggy_df.isnull().sum()}")
print(f"Zomato:\n{zomato_df.isnull().sum()}")
print(f"Indian Restaurants:\n{indian_restaurants_df.isnull().sum()}")

swiggy_df['source'] = 'Swiggy' 
zomato_df['source'] = 'Zomato'
indian_restaurants_df['source'] = 'Indian Restaurants'

# merged_df = pd.merge(swiggy_df, zomato_df, left_on=['Name','Area'], right_on=['name','location'], how='inner')
df = pd.concat([swiggy_df, zomato_df, indian_restaurants_df], axis=0)
df

print(f"{df.isnull().sum()}")

df['rating'] = df['rating'].replace('--', pd.NA)
df = df.drop_duplicates(subset=['name', 'location'])
df = df.drop_duplicates(subset=['name', 'address'])
df['cost_for_two'] = df['cost_for_two'].str.replace('₹', '').str.strip()
df['rating'] = pd.to_numeric(df['rating'], errors='coerce')
df['rating_count'] = pd.to_numeric(df['rating_count'], errors='coerce')
df['cost_for_two'] = pd.to_numeric(df['cost_for_two'], errors='coerce')

# Since all columns are required for the recommendation system, we will drop rows with any null values 
df = df.dropna()
df

df.to_csv('all_data/merged_restaurant_data.csv',index=False)
df.to_csv('Data/merged_restaurant_data.csv',index=False)
