==> Scripts/5_data_merging.ipynb <==
import pandas as pd
import os

df = pd.read_csv('../Data/feature_engineered_data.csv')
df_ratings = pd.read_csv('../Data/merged_restaurant_data.csv')

df

df = df[df['success'] == True]
df.drop_duplicates(subset=['place_id'], inplace=True)
df.reset_index(drop=True, inplace=True)
df.drop(columns=['success','index'], inplace=True)
df

# combine the two dataframes on name and address
df = pd.merge(df, df_ratings, on=['name', 'address'], how='inner')
df

df.to_csv('../Data/final_df.csv', index=False)
